requirements.txt
aenum==2.2.4
attrs==19.3.0
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
enum34==1.1.10
Fiona==1.8.13.post1
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
geopandas==0.6.3
gunicorn==19.10.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
munch==2.5.0
numpy==1.16.6
pandas==0.24.2
psycopg2-binary==2.8.5
pyproj==2.2.2
python-dateutil==2.8.1
pytz==2020.1
Shapely==1.7.0
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
